docs: add general guidance for contributors
Documents were added with guidelines and rules for all people interested
in contributing and being members of a community as authors, reviewers
or maintainers.

Includes:
- project scope and objectives;
- pattern for commit messages;
- PR review process rules;
- and more.

Furthermore, it will be used as a draft of rules and best practices to
be used for any new project.

Also, some static code analysis issues are resolved.

Signed-off-by: Claudio André <claudioandre.br@gmail.com>