Docker: sign the Docker image using Sigstore
This is NOT going to drastically improve things, but it seems like good
practice.

Sigstore is a new standard for signing, verifying, and protecting software.

Sigstore enables developers to validate that the software they are using is
exactly what it claims to be using cryptographic digital signatures and
transparency log technologies.

Signed-off-by: Claudio André <claudioandre.br@gmail.com>