feat(ci): use a VM to build flatpak bundle (#307)
In order to be able to test the package, we need to build the image
outside of a Docker container. Both snap and flatpak cannot run
correctly inside a Docker container.

Our infrastructure at Gitlab is focused on Docker.
So add a VM that allows us to create and test the flatpak bundle.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>