maint: avoid downloading `helper.sh` when possible
In the case of snap this is not possible, but in other cases it is possible.

This is a repository that contains a toolbox to solve the problem of building
and testing in multiple environments. These environments can be very
heterogeneous and even without network access. Therefore, some hacks
may be necessary for us to create something that works in such heterogeneous
environments.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>