maint: add provenance to GitHub Releases
Implement the level 1 requirements of the SLSA framework. Better than
nothing.

Supply-chain Levels for Software Artifacts, or SLSA ("salsa"):
It’s a security framework, a checklist of standards and controls to
prevent tampering, improve integrity, and secure packages and infrastructure.
It’s how you get from "safe enough" to being as resilient as possible,
at any link in the chain.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>