deploy: make builds repeatable and reproducible
All package build must "lock" and get the same commit in order to
make the process of creating a package repeatable. This will force
us to change the build commit hash every time a new build occurs.

For dev packages, a temporary change will occur in the 'release' branch
whenever a new package is released. In the end, for "real releases", this
will mean committing the final commit hash to 'main'.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>