maint: change the frequency of dependency updates
Set updates to weekly to make repository management easier.
Management in the future may become much more complicated but this
change reduces the pressure on maintainers.

Also, remove the "Docker" type check performed in the root folder: there
is no Dockerfile in the main project folder.

Finally, as the ASAN/UBSAN verification runs smoothly now, there is no
need to tag me in the PR.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>