IDE: adjust the Docker env used by devcontainer
The base Docker image generated and made available by Microsoft [a]
already has an ubuntu user (UID=1000) and a vscode user (UID=1001)
registered. No problem occurs if we directly run a container based
on this image, even selecting vscode as user, but for some reason,
when someone uses the VS Code feature [b], 'vscode-server' creates
a container based on the image [a] but it defines the user who runs
the container on the host as UID=1001 (a non-existent user on my host).

This results in access denied issues on any file edited using VS Code.
Therefore, we need to at least create a workaround, forcing the vscode
user to be UID=1000.

[a] mcr.microsoft.com/vscode/devcontainers/base:noble
[b] The Dev Containers extension lets you use a Docker container as
    a full-featured development environment.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>