flatpak: add support for OpenCL (HPC)
It is possible to use GPU acceleration via OpenCL within flatpak:
- https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1535
- https://github.com/flathub/com.openwall.John/pull/13#issuecomment-2341867599

However, it's not clear what is actually needed for this to work (poor
documentation? or I'm failing to find it?).
- what is the minimum required to run `cl-info` inside a flatpak?

HPC (high performance computing) using GPU is very important to us.
NVIDIA is the target at the moment.

Our community does not have suitable hardware with CUDA installed to
use/test this PR.

=> Note: Review/Revert/Improve:
. Should I also add org.freedesktop.Platform.HIP.Nvidia?
. Is the `add-extensions:` necessary or useful? I want to enable NVIDIA support
but allow poor people (like me) to continue using flatpak, e.g., for CPU
formats.
. Test when possible:
```
+        "--socket=x11",
+        "--device=all"
```

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>