docs: add notes about the testing packages
We need to make it clear that the GitHub release called bleeding is NOT
exactly the one we use as a testing environment, i.e. the `bleeding`
GitHub release MAY be out of date.

The test environment that CAN be updated with upstream source code
is the Docker image and the Snap package.

See also https://github.com/openwall/john-packages/blob/main/docs/release-process.md#the-bleeding-release.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>