refactor(merge-bot): restructure code (#612)
It's just a reorganization of the code with no new features included.

Well, actually, the trial mode is something new, but not a new feature.
- to run the bot without the PR being merged, just use the text
  (without the quotation marks): "bot: MERGE trial".

The list of accepted commands is:

To call the bot, run all the validations and merge the PR, type:
bot: MERGE [optional comments]

To call the bot, bypass the validations and merge the PR (GitHub rules
still apply), type:
bot: MERGE skip

To call the bot without the PR being merged at the end, type:
bot: MERGE trial

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>