feat(merge-bot): call the bot via PR comment (#625)
Allow the merge bot to be called via a comment in a PR.

In the cloud, things are eventually consistent, so after a change to
a PR, the user must wait a few seconds before calling the merge bot,
otherwise the PR status will be unstable or unknown.

For this reason, calling the merge bot via comments on a PR review
is a bad idea. In this situation, the PR itself is being changed.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>