security(merge-bot): allow app update workflow (#629)
Allow the bot to update workflow files (this is necessary because we
will use it for dependabot updates, which impact workflows a lot).

Human intervention is required to start the bot.

Maybe that's why I was seeing strange errors. GitHub has (just now)
added an error message that didn't exist before.

Error message :
- refusing to allow a GitHub App to create or update workflow
  `.github/workflows/approve-it.yml` without `workflows` permission.