chore(trivy): temporarily remove secrets check (#637)
Revert this commit in a week or so.

As it stands now, Trivy isn't evaluating the source code, it's evaluating
the cache which doesn't reflect the current situation of the tree.

I'm trying to force it to "notice that the source code has changed".

- Quite interesting (the same workflow executed)!
Trivy fs / scan-fs (push) Failing after 19s
Trivy fs / scan-fs (pull_request) Successful in 24s

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>