ci: fail if formatting is recommended (#660)
Ask Linter to reformat the source code and fail if Linter suggests
changes. This means that the code should have been formatted differently
for the PR.

Something like this has already happened to me: the Linter failed
because of a single/double quotation mark when a double/single was
expected, then, I didn't receive any information and I didn't find
out what happened.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>