docs: alert about known release errors (#709)
There are relevant bugs in previous versions of `john`. It's really
important that people know this and stop using older versions of the
software.
This is a temporary warning, which should be removed as soon as we
have a better mechanism for releasing improvements to the software.

People who use older versions may simply be wasting their time by doing
so. As a result of the bugs detected and fixed in the main repository,
a new release is more than welcome.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>