docs: tune the process of creating a release (#714)
In order for the run/Defaults file to contain the final commit hash
(which has not yet been merged in main), it must not be changed later
after the binaries have been built.

It is therefore necessary to run the CI to obtain information on the
version of the software used in it, update the commit and then perform
the actual/final build.

The CI will be run twice:
- first to get the version to update the documentation;
  - update the content of the commit with the correct text and push.
- then run the actual build of the package.
  - the commit hash saved in build logs will not be changed (if
    changes are not submitted after this point).

Once this is done, just use a fast-forward merge.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>